refactor(platform/api): relocate auth into domains/auth/service, fix stale keystore docstring
